Add CourseMembership
This class will be used to store information about what courses does a
person belong to; it is LMS's responsibilty to present the user with
option to enroll only in those courses that they are allowed to
(i.e. CourseMembership is only stating a fact of user being a member of
some course, regardless of whether they should be allowed to enroll in
that course or not).