Enhance UI
Quite a few things are done here. One is that UI now encompasses
changes that the client wanted; it's still far from pretty, but is
at least readable (on my laptop...)

The next thing is removal of some of the unnecessary CSS code. But I've
also introduced a bad practice here: moved from having nice and specific
header tags to the general <p> tag with different classes. I know this
has to be rewritten later, but right now it was the fastest way to
get the styling that I want.

I also removed some stuff that was responsible for adaptability as,
let's be honest, this code is already very far from providing flexible
UI, and to make it truly available I need to read something on
interfaces and CSS first.